requirements.txt
django==4.1
djangorestframework==3.13.1

mysqlclient==2.0.1

# -------------------------------- Dev

ipdb
